RavenDB-12389 Moving classes that implements IOperationProgress and IOperationResult and are exposed in client operations to Raven.Client project
