+++ src/Raven.Client/ServerWide/Operations/CompactionProgress.cs
@@ -2,7 +2,7 @@
 using Raven.Client.Documents.Operations;
 using Sparrow.Json.Parsing;
 
-namespace Raven.Server.Documents
+namespace Raven.Client.ServerWide.Operations
 {
     public class CompactionProgressBase
     {
