+++ src/Raven.Client/ServerWide/Operations/RestoreResult.cs
@@ -1,7 +1,7 @@
 using Raven.Client.Documents.Smuggler;
 using Sparrow.Json.Parsing;
 
-namespace Raven.Server.Documents.PeriodicBackup.Restore
+namespace Raven.Client.ServerWide.Operations
 {
     public class RestoreResult : SmugglerResult
     {
